pyproject.toml
[tool.poetry]
name = "grounded-language"
version = "0.1.0"
description = ""
authors = ["Henry Prior <henry.g.prior@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
jax = "^0.2.8"
dm_fast_mapping = { git = 'https://github.com/deepmind/dm_fast_mapping.git', branch = 'master' }
dm-haiku = { git = "https://github.com/deepmind/dm-haiku.git", branch = 'master' }
pygame = "^2.0.1"
jaxlib = "^0.1.59"
tokenizers = "^0.10.0"
flax = "^0.3.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
reorder-python-imports = "^2.3.6"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
